Move gator test resources into pkg/gator/fixtures
